fix type of variable gram fix name of variable specialty

+++ src/main/java/com/pfizer/sacchon/model/Carb.java
@@ -12,7 +12,7 @@ public class Carb {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     private long id;
-    private long gram;
+    private double gram;
     private Date date;
     @ManyToOne
     @JoinColumn (name = "patient_id")
